# src/main/resources/api-docs-01.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:4444","description":"Generated server url"}],"tags":[{"name":"Authentication API","description":"Returns JWT Token"},{"name":"Prescriber API","description":"Operations related to prescribers and their assigned patients and bags"}],"paths":{"/auth/login":{"post":{"tags":["Authentication API"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/prescriber/retrieve/prescriber":{"post":{"tags":["Prescriber API"],"summary":"Retrieve prescriber by ID","description":"Fetches full details for a prescriber by their ID","operationId":"retrievePrescriber","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrescriberIdRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/prescriber/retrieve/patient":{"post":{"tags":["Prescriber API"],"summary":"Retrieve patient by ID","description":"Fetches full details for a patient by their ID","operationId":"retrievePatient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientIdRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/prescriber/patients/list":{"post":{"tags":["Prescriber API"],"summary":"List all assigned patients","description":"Returns all patients assigned to the authenticated (By JWT) prescriber","operationId":"viewMyPatients","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/prescriber/patient/prescriptions/history":{"post":{"tags":["Prescriber API"],"summary":"Get patient prescription history","description":"Returns all previous prescriptions for a patient","operationId":"getPatientPrescriptionHistory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientIdRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/prescriber/patient/bag/view":{"post":{"tags":["Prescriber API"],"summary":"View patient and bag","description":"Returns patient and associated bag details","operationId":"viewPatientAndBag","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientIdRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/prescriber/patient/assign":{"post":{"tags":["Prescriber API"],"summary":"Assign patient to prescriber","description":"Links a patient with a specific prescriber","operationId":"assignPrescriber","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignPrescriber"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/prescriber/create":{"post":{"tags":["Prescriber API"],"summary":"Create a new prescriber","description":"Registers a new prescriber using their details","operationId":"createNewPrescriber","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrescriberRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/prescriber/bag/state":{"post":{"tags":["Prescriber API"],"summary":"Get bag state","description":"Returns the current state of a bag","operationId":"getBagState","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientIdRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/prescriber/bag/set/state":{"post":{"tags":["Prescriber API"],"summary":"Set bag state","description":"Sets the state of a bag (e.g., SEALED, UNSEALED)Quite powerful as it overrides any state and could break machine logic","operationId":"setBagState","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetBagState"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/prescriber/bag/prescription/update":{"post":{"tags":["Prescriber API"],"summary":"Update prescription","description":"Modifies the prescription for a specific bag","operationId":"changePrescription","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePrescription"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/prescriber/bag/create":{"post":{"tags":["Prescriber API"],"summary":"Create new bag","description":"Creates a new medication bag for a patient","operationId":"createNewBag","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BagCreation"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/patient/upload/photo/path":{"post":{"tags":["Patient API"],"summary":"Upload patient photo path","description":"Stores the path of a patient's photo in the system","operationId":"uploadPhotoPath","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoUploadRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/patient/retrieve/patient":{"post":{"tags":["Patient API"],"summary":"Retrieve patient by ID","description":"Returns full patient details using patient ID","operationId":"retrievePatient_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientIdRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/patient/patient/bag/view":{"post":{"tags":["Patient API"],"summary":"View patient and their bag","description":"Fetches both the patient's details and their associated medication bag","operationId":"viewPatientAndBag_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientIdRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/patient/create":{"post":{"tags":["Patient API"],"summary":"Create a new patient","description":"Registers a new patient. Date of birth must be in format dd-MM-yyyy. Email must be unique.","operationId":"createNewPatient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/device/retrieve/prescriber":{"post":{"tags":["Device API"],"summary":"Retrieve prescriber information","description":"Returns detailed information of a prescriber using their ID","operationId":"retrievePrescriber_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrescriberIdRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/device/retrieve/patient":{"post":{"tags":["Device API"],"summary":"Retrieve patient information","description":"Fetches the complete details of a patient by their ID","operationId":"retrievePatient_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientIdRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/device/get/bag/state":{"post":{"tags":["Device API"],"summary":"Get bag state for a patient","description":"Retrieves the current state of the medication bag assigned to a specific patient by their ID","operationId":"getBagState_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientIdRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/device/bag/set/state":{"post":{"tags":["Device API"],"summary":"Set new state for a bag","description":"MUST be in UPPERCASE. Updates the state of a medication bag (e.g., UNSEALED, SEALED, LOADED)Quite powerful as it overrides any state and could break the machine logic","operationId":"setBagState_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetBagState"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/prescriber/patient/unsealed/bags":{"get":{"tags":["Prescriber API"],"summary":"Get UNSEALED bags only","description":"Retrieves bags in UNSEALED state only for a prescriber's patients","operationId":"retrieveAllUnsealedBags","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BagEntity"}}}}}}}},"/api/prescriber/patient/loaded/bags":{"get":{"tags":["Prescriber API"],"summary":"Get LOADED bags only","description":"Retrieves bags in LOADED state only for a prescriber's patients","operationId":"retrieveUnsealedBags","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BagEntity"}}}}}}}},"/api/prescriber/patient/bags":{"get":{"tags":["Prescriber API"],"summary":"Get all bags for prescriber's patients","description":"Retrieves all bags associated with a prescriber's patients","operationId":"retrieveAllBags","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BagEntity"}}}}}}}}},"components":{"schemas":{"AuthenticationRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"AuthenticationResponse":{"required":["token"],"type":"object","properties":{"token":{"type":"string"}}},"PrescriberIdRequest":{"required":["prescriberId"],"type":"object","properties":{"prescriberId":{"type":"string","format":"uuid"}}},"PatientIdRequest":{"required":["patientId"],"type":"object","properties":{"patientId":{"type":"string","format":"uuid"}}},"AssignPrescriber":{"required":["patientId","prescriberId"],"type":"object","properties":{"patientId":{"type":"string","format":"uuid"},"prescriberId":{"type":"string","format":"uuid"}}},"PrescriberRequest":{"required":["email","name","password"],"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"SetBagState":{"required":["bagState","patientId","prescriberId"],"type":"object","properties":{"patientId":{"type":"string","format":"uuid"},"prescriberId":{"type":"string","format":"uuid"},"bagState":{"type":"string","enum":["UNSEALED","SEALED","LOADED","DISPENSED","DISCARDED"]}}},"UpdatePrescription":{"required":["patientId","prescription"],"type":"object","properties":{"patientId":{"type":"string","format":"uuid"},"prescription":{"type":"string"}}},"BagCreation":{"required":["patientId","prescription"],"type":"object","properties":{"patientId":{"type":"string","format":"uuid"},"prescription":{"type":"string"}}},"PhotoUploadRequest":{"required":["patientId","photoPath"],"type":"object","properties":{"patientId":{"type":"string","format":"uuid"},"photoPath":{"type":"string"}}},"PatientRequest":{"required":["dob","email","name","password","phone"],"type":"object","properties":{"prescriberId":{"type":"string","format":"uuid"},"name":{"type":"string"},"dob":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"password":{"type":"string"},"faceImagePath":{"type":"string"}}},"BagEntity":{"required":["createdAt","prescription","state"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"patientId":{"type":"string","format":"uuid"},"prescription":{"type":"string"},"state":{"type":"string","enum":["UNSEALED","SEALED","LOADED","DISPENSED","DISCARDED"]},"createdAt":{"type":"string","format":"date-time"}}}}}}